fix(clippy): 同型 u32 -> u32 キャスト除去
helpers::last_digit() は u32 を返すので `last as u32` は不要 cast、
直前 commit で contains(&(last as u32)) に書き換えた際に残った。
clippy::unnecessary_cast lint で fail していたので contains(&last) に修正。

+++ crates/furigana/src/scoring/numbers.rs
@@ -279,7 +279,7 @@ fn scale_trailing_counter_suffix(
     let last = crate::numbers::helpers::last_digit(&normalized);
     // last_digit が rule に match すればその suffix を、 無ければ default を採用
     for r in &rule.rules {
-        if r.last_digit.contains(&(last as u32)) {
+        if r.last_digit.contains(&last) {
             return Some(r.suffix.clone());
         }
     }
